# schemas/openapi.yaml
openapi: 3.1.0
info:
  title: Agent OS Control Plane API
  description: |
    API for the Agent OS Control Plane — vendor-neutral orchestration,
    governance, and observability for AI agents.
  version: 0.1.0
  contact:
    name: Agent OS Team
    email: api@agent-os.dev
  license:
    name: Proprietary
    url: https://agent-os.dev/license
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
  - url: https://api.agent-os.dev/v1
    description: Production
security:
  - bearerAuth: []
paths:
...
  /agents/{agent_id}/run:
    post:
      operationId: createRun
      summary: Create and start a run
      tags: [Runs]
      parameters:
        - $ref: '#/components/parameters/AgentIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunCreate'
      responses:
        '202':
          description: Run accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Run'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /runs:
    get:
      operationId: listRuns
      summary: List all runs
      tags: [Runs]
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
        - name: agent_id
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, running, completed, failed, cancelled]
      responses:
        '200':
          description: Runs listed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunList'
  /runs/{run_id}:
    get:
      operationId: getRun
      summary: Get run by ID
      tags: [Runs]
      parameters:
        - name: run_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Run found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Run'
        '404':
          $ref: '#/components/responses/NotFound'
  /runs/{run_id}/cancel:
    post:
      operationId: cancelRun
      summary: Cancel a running run
      tags: [Runs]
      parameters:
        - name: run_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Run cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Run'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
    OffsetParam:
      name: offset
      in: query
      schema:
        type: integer
        default: 0
        minimum: 0
    AgentIdParam:
      name: agent_id
      in: path
      required: true
      schema:
        type: string
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Run:
      type: object
      properties:
        id:
          type: string
        agent_id:
          type: string
        status:
          type: string
          enum: [pending, running, completed, failed, cancelled]
        prompt:
          type: string
        context:
          type: object
        result:
          type: object
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
    RunCreate:
      type: object
      required: [prompt]
      properties:
        prompt:
          type: string
        context:
          type: object
        options:
          type: object
          properties:
            stream:
              type: boolean
              default: false
            timeout:
              type: integer
              default: 30
    RunList:
      type: object
      properties:
        runs:
          type: array
          items:
            $ref: '#/components/schemas/Run'
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
...
    Error:
      type: object
      properties:
        detail:
          type: string
        code:
          type: string